backend/requirements.txt
# ─── Backend Dependencies ─────────────────────────────────────────────────────
# Updated for Python 3.14 compatibility

fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.30.0  # PostgreSQL async driver
aiosqlite==0.20.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.4
httpx==0.27.0
psycopg2-binary==2.9.9  # for sync ETL worker
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1
python-dateutil==2.9.0.post0
orjson==3.10.3
structlog==24.1.0
prometheus-fastapi-instrumentator==7.0.0
tenacity==8.3.0
email-validator==2.1.1
pyyaml>=6.0

# ─── Data & Analytics ────────────────────────────────────────────────────────
# Updated versions for Python 3.14 compatibility
duckdb>=1.0.0
polars>=1.0.0
numpy>=2.0.0
